separating list parsing into utility
